Implement compiled ZXing core, UPC_EAN_EXTENSION
Specially modified version of ZXing that supports scanning and
returning the UPC_EAN_EXTENSION when combined with standard barcodes.
This is necessary to allow the scanning of comic books, which contain
issue information in the extension. Extensions are often used for the
issue information with magazines and pricing information with books,
whereas comic books use the book barcode with the issue information
extension.
- Modification taken from the source of Android Buddy
(http://code.google.com/p/cupcake-frosting/source/browse)