Disable the default location bar and add cordova instance
The cordova implementation is commonly lowercase, so this accounts for
the possibility that it is configured that way.
The location bar still displays with a null value, as it would with the
default as true, but unlike the previous setting, this seems to make
the page actually display as well.